Notification part

+++ server/src/main/java/com/Cranco/Cranco/Notification/NotificationRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.neo4j.repository.config.EnableNeo4jRepositories;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NotificationRepository extends Neo4jRepository<Notification, Integer> {
+public interface NotificationRepository extends Neo4jRepository<Notification, Long> {
     List<Notification> findByReceiverAndSeen(User receiver, boolean seen);
